# docs/api/openapi3.yaml
openapi: 3.0.1
info:
  title: sync
  description: ""
  version: 0.0.1
servers:
- url: ""
...
paths:
...
  /comments/{commentId}:
    delete:
      tags:
      - comment
      summary: Delete Comment
      description: Delete Comment
      operationId: DeleteComment
      parameters:
      - name: commentId
        in: path
        description: Comment ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - comment
      summary: Update Comment
      description: Update Comment
      operationId: UpdateComment
      parameters:
      - name: commentId
        in: path
        description: Comment ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentRequest"
            examples:
              UpdateComment:
                value: "{\"content\":\"Updated comment content\"}"
      responses:
        "204":
          description: "204"
  /comments/{commentId}/accept:
    put:
      tags:
      - comment
      summary: Accept Comment
      description: Accept Comment
      operationId: AcceptComment
      parameters:
      - name: commentId
        in: path
        description: Comment ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    delete:
      tags:
      - comment
      summary: Unaccept Comment
      description: Unaccept Comment
      operationId: UnacceptComment
      parameters:
      - name: commentId
        in: path
        description: Comment ID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
...
  /posts/{slug}/comments:
    get:
      tags:
      - comment
      summary: Get Post Comments
      description: Get Post Comments
      operationId: GetPostComments
      parameters:
      - name: slug
        in: path
        description: Post Slug
        required: true
        schema:
          type: string
      - name: first
        in: query
        description: Number of items to return from the beginning of the list
        required: false
        schema:
          type: string
      - name: after
        in: query
        description: "Cursor for pagination, indicating the position after which to\
          \ return items"
        required: false
        schema:
          type: string
      - name: last
        in: query
        description: Number of items to return from the end of the list
        required: false
        schema:
          type: string
      - name: before
        in: query
        description: "Cursor for pagination, indicating the position before which\
          \ to return items"
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCommentsResponse"
              examples:
                GetPostComments:
                  value: "{\"comments\":{\"pageInfo\":{\"size\":1,\"hasNextPage\"\
                    :false,\"hasPreviousPage\":false,\"startCursor\":\"start-cursor\"\
                    ,\"endCursor\":\"end-cursor\"},\"nodes\":[{\"cursor\":\"cursor\"\
                    ,\"content\":{\"id\":1,\"author\":{\"handle\":\"john-doe\",\"\
                    name\":\"John Doe\",\"profileImageUrl\":\"https://example.com/john.png\"\
                    },\"isPostAuthor\":false,\"content\":\"Comment content\",\"isDeleted\"\
                    :false,\"isAccepted\":false,\"createdAt\":\"2026-01-01T00:00:00Z\"\
                    ,\"updatedAt\":\"2026-01-01T00:00:00Z\"}}]}}"
    post:
      tags:
      - comment
      summary: Create Comment
      description: Create Comment
      operationId: CreateComment
      parameters:
      - name: slug
        in: path
        description: Post Slug
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
            examples:
              CreateComment:
                value: "{\"content\":\"Comment content\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCommentResponse"
              examples:
                CreateComment:
                  value: "{\"id\":1}"
...
  /projects/{handle}/tags:
    get:
      tags:
      - tag
      summary: Get Project Tags
      description: 프로젝트의 태그 목록을 조회합니다.
      operationId: GetProjectTags
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTagsResponse"
              examples:
                GetProjectTags:
                  value: "{\"tags\":[{\"id\":1,\"name\":\"java\",\"description\":\"\
                    자바 관련 태그\",\"postCount\":10,\"followerCount\":3,\"projectHandle\"\
                    :null,\"isFollowing\":false},{\"id\":2,\"name\":\"spring\",\"\
                    description\":\"스프링 프레임워크 관련 태그\",\"postCount\":5,\"followerCount\"\
                    :1,\"projectHandle\":null,\"isFollowing\":true}]}"
    post:
      tags:
      - tag
      summary: Create Project Tag
      description: 프로젝트 태그를 생성합니다. 프로젝트 관리자만 접근할 수 있습니다.
      operationId: CreateProjectTag
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
            examples:
              CreateProjectTag:
                value: "{\"name\":\"java\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTagResponse"
              examples:
                CreateProjectTag:
                  value: "{\"id\":1,\"name\":\"java\",\"description\":\"\",\"postCount\"\
                    :0}"
...
  /projects/{handle}/tags/unverified:
    get:
      tags:
      - tag
      summary: Get Project Unverified Tags
      description: 프로젝트의 인증되지 않은 태그 목록을 조회합니다. 프로젝트 관리자만 접근할 수 있습니다.
      operationId: GetProjectUnverifiedTags
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTagsResponse"
              examples:
                GetProjectUnverifiedTags:
                  value: "{\"tags\":[{\"id\":1,\"name\":\"java\",\"description\":\"\
                    자바 관련 태그\",\"postCount\":10,\"followerCount\":3,\"projectHandle\"\
                    :null,\"isFollowing\":false},{\"id\":2,\"name\":\"spring\",\"\
                    description\":\"스프링 프레임워크 관련 태그\",\"postCount\":5,\"followerCount\"\
                    :1,\"projectHandle\":null,\"isFollowing\":true}]}"
  /projects/{handle}/tags/{name}:
    delete:
      tags:
      - tag
      summary: Reject Project Tag
      description: 프로젝트 태그를 삭제합니다. 프로젝트 관리자만 접근할 수 있습니다.
      operationId: RejectProjectTag
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      - name: name
        in: path
        description: 삭제할 태그 이름
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - tag
      summary: Update Project Tag
      description: 프로젝트 태그 정보를 수정합니다. 프로젝트 관리자만 접근할 수 있습니다.
      operationId: UpdateProjectTag
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      - name: name
        in: path
        description: 수정할 태그의 현재 이름
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/UpdateTagRequest"
            examples:
              UpdateProjectTag:
                value: "{\"name\":\"kotlin\",\"description\":\"코틀린 관련 태그\"}"
      responses:
        "204":
          description: "204"
...
  /projects/{handle}/tags/{name}/verify:
    patch:
      tags:
      - tag
      summary: Verify Project Tag
      description: 프로젝트 태그를 인증합니다. 프로젝트 관리자만 접근할 수 있습니다.
      operationId: VerifyProjectTag
      parameters:
      - name: handle
        in: path
        description: 프로젝트 핸들
        required: true
        schema:
          type: string
      - name: name
        in: path
        description: 인증할 태그 이름
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
...
  /search/tags:
    get:
      tags:
      - tag
      summary: Search Tags
      description: 검색어로 태그를 검색합니다.
      operationId: SearchTags
      parameters:
      - name: handle
        in: query
        description: 프로젝트 핸들 (지정 시 프로젝트 범위로 검색)
        required: false
        schema:
          type: string
      - name: query
        in: query
        description: 태그 검색어
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTagsResponse"
              examples:
                SearchTags:
                  value: "{\"tags\":[{\"id\":1,\"name\":\"java\",\"description\":\"\
                    자바 관련 태그\",\"postCount\":10,\"followerCount\":3,\"projectHandle\"\
                    :null,\"isFollowing\":false},{\"id\":2,\"name\":\"spring\",\"\
                    description\":\"스프링 프레임워크 관련 태그\",\"postCount\":5,\"followerCount\"\
                    :1,\"projectHandle\":null,\"isFollowing\":true}]}"
...
  /tags:
    get:
      tags:
      - tag
      summary: Get All Tags
      description: 인증된 전역 태그 전체 목록을 페이지 단위로 조회합니다.
      operationId: GetAllTags
      parameters:
      - name: page
        in: query
        description: Page (Integer)
        required: false
        schema:
          type: string
      - name: size
        in: query
        description: Page Size (Integer)
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllTagsResponse"
              examples:
                GetAllTags:
                  value: "{\"tags\":{\"pageInfo\":{\"page\":1,\"size\":1,\"hasNextPage\"\
                    :true,\"hasPreviousPage\":true},\"content\":[{\"id\":1,\"name\"\
                    :\"java\",\"description\":\"자바 관련 태그\",\"postCount\":10,\"followerCount\"\
                    :3,\"projectHandle\":null,\"isFollowing\":false},{\"id\":2,\"\
                    name\":\"spring\",\"description\":\"스프링 프레임워크 관련 태그\",\"postCount\"\
                    :5,\"followerCount\":1,\"projectHandle\":null,\"isFollowing\"\
                    :true}]}}"
    post:
      tags:
      - tag
      summary: Create Tag
      description: 전역 태그를 생성합니다. 관리자만 접근할 수 있습니다.
      operationId: CreateTag
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
            examples:
              CreateTag:
                value: "{\"name\":\"java\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTagResponse"
              examples:
                CreateTag:
                  value: "{\"id\":1,\"name\":\"java\",\"description\":\"\",\"postCount\"\
                    :0}"
  /tags/merge:
    post:
      tags:
      - tag
      summary: Merge Tags
      description: 동일한 풀에 속한 두 태그를 하나로 병합합니다. 관리자만 접근할 수 있습니다.
      operationId: MergeTags
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/MergeTagsRequest"
            examples:
              MergeTags:
                value: "{\"sourceTagId\":1,\"targetTagId\":2}"
      responses:
        "204":
          description: "204"
...
  /tags/{id}:
    get:
      tags:
      - tag
      summary: Get Tag
      description: 태그 ID로 태그 상세 정보를 조회합니다. 프로젝트 태그는 프로젝트가 공개이거나 요청자가 프로젝트 팀원인 경우에만
        조회할 수 있습니다.
      operationId: GetTag
      parameters:
      - name: id
        in: path
        description: 태그 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTagResponse"
              examples:
                GetTag:
                  value: "{\"tag\":{\"id\":1,\"name\":\"java\",\"description\":\"자\
                    바 관련 태그\",\"postCount\":10,\"followerCount\":3,\"projectHandle\"\
                    :null,\"isFollowing\":false}}"
  /tags/{name}:
    delete:
      tags:
      - tag
      summary: Reject Tag
      description: 전역 태그를 삭제합니다. 관리자만 접근할 수 있습니다.
      operationId: RejectTag
      parameters:
      - name: name
        in: path
        description: 삭제할 태그 이름
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
    patch:
      tags:
      - tag
      summary: Update Tag
      description: 전역 태그 정보를 수정합니다. 관리자만 접근할 수 있습니다.
      operationId: UpdateTag
      parameters:
      - name: name
        in: path
        description: 수정할 태그의 현재 이름
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/UpdateTagRequest"
            examples:
              UpdateTag:
                value: "{\"name\":\"kotlin\",\"description\":\"코틀린 관련 태그\"}"
      responses:
        "204":
          description: "204"
  /tags/{name}/verify:
    patch:
      tags:
      - tag
      summary: Verify Tag
      description: 전역 태그를 인증합니다. 관리자만 접근할 수 있습니다.
      operationId: VerifyTag
      parameters:
      - name: name
        in: path
        description: 인증할 태그 이름
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
...
components:
  schemas:
...
    GetTagResponse:
      title: GetTagResponse
      type: object
      properties:
        tag:
          required:
          - description
          - followerCount
          - id
          - isFollowing
          - name
          - postCount
          type: object
          properties:
            isFollowing:
              type: boolean
              description: 요청자가 해당 태그를 팔로우하고 있는지 여부 (프로젝트 태그는 항상 false)
            name:
              type: string
              description: 태그 이름
            postCount:
              type: number
              description: 태그가 사용된 게시물 수
            description:
              type: string
              description: 태그 설명
            id:
              type: number
              description: 태그 ID
            followerCount:
              type: number
              description: 태그를 팔로우하는 사용자 수
            projectHandle:
              type: string
              description: 프로젝트 태그인 경우 해당 프로젝트의 핸들 (전역 태그인 경우 없음)
              nullable: true
...
    UpdateCommentRequest:
      title: UpdateCommentRequest
      required:
      - content
      type: object
      properties:
        content:
          type: string
          description: Comment content
    CreateTagResponse:
      title: CreateTagResponse
      required:
      - description
      - id
      - name
      - postCount
      type: object
      properties:
        name:
          type: string
          description: 태그 이름
        postCount:
          type: number
          description: 태그가 사용된 게시물 수
        description:
          type: string
          description: 태그 설명
        id:
          type: number
          description: 생성된 태그 ID
...
    GetAllTagsResponse:
      title: GetAllTagsResponse
      type: object
      properties:
        tags:
          required:
          - content
          - pageInfo
          type: object
          properties:
            pageInfo:
              required:
              - hasNextPage
              - hasPreviousPage
              - page
              - size
              type: object
              properties:
                size:
                  type: number
                  description: Page Size
                hasNextPage:
                  type: boolean
                  description: Has Next Page
                hasPreviousPage:
                  type: boolean
                  description: Has Previous Page
                page:
                  type: number
                  description: Page Number
              description: Page Info
            content:
              type: array
              description: Content
              items:
                required:
                - description
                - followerCount
                - id
                - isFollowing
                - name
                - postCount
                type: object
                properties:
                  isFollowing:
                    type: boolean
                    description: 요청자가 해당 태그를 팔로우하고 있는지 여부 (프로젝트 태그는 항상 false)
                  name:
                    type: string
                    description: 태그 이름
                  postCount:
                    type: number
                    description: 태그가 사용된 게시물 수
                  description:
                    type: string
                    description: 태그 설명
                  id:
                    type: number
                    description: 태그 ID
                  followerCount:
                    type: number
                    description: 태그를 팔로우하는 사용자 수
                  projectHandle:
                    type: string
                    description: 프로젝트 태그인 경우 해당 프로젝트의 핸들 (전역 태그인 경우 없음)
                    nullable: true
...
    MergeTagsRequest:
      title: MergeTagsRequest
      type: object
...
    GetTagsResponse:
      title: GetTagsResponse
      required:
      - tags
      type: object
      properties:
        tags:
          type: array
          description: 태그 목록
          items:
            required:
            - description
            - followerCount
            - id
            - isFollowing
            - name
            - postCount
            type: object
            properties:
              isFollowing:
                type: boolean
                description: 요청자가 해당 태그를 팔로우하고 있는지 여부 (프로젝트 태그는 항상 false)
              name:
                type: string
                description: 태그 이름
              postCount:
                type: number
                description: 태그가 사용된 게시물 수
              description:
                type: string
                description: 태그 설명
              id:
                type: number
                description: 태그 ID
              followerCount:
                type: number
                description: 태그를 팔로우하는 사용자 수
              projectHandle:
                type: string
                description: 프로젝트 태그인 경우 해당 프로젝트의 핸들 (전역 태그인 경우 없음)
                nullable: true
...
    UpdateTagRequest:
      title: UpdateTagRequest
      required:
      - name
      type: object
      properties:
        name:
          type: string
          description: 변경할 태그 이름
        description:
          type: string
          description: 변경할 태그 설명
          nullable: true
...
    CreateTagRequest:
      title: CreateTagRequest
      required:
      - name
      type: object
      properties:
        name:
          type: string
          description: 태그 이름
...
    GetCommentsResponse:
      title: GetCommentsResponse
      required:
      - comments
      type: object
      properties:
        comments:
          required:
          - nodes
          - pageInfo
          type: object
          properties:
            nodes:
              type: array
              description: Nodes
              items:
                required:
                - content
                - cursor
                type: object
                properties:
                  cursor:
                    type: string
                    description: Node Cursor
                  content:
                    required:
                    - author
                    - createdAt
                    - id
                    - isAccepted
                    - isDeleted
                    - isPostAuthor
                    - updatedAt
                    type: object
                    properties:
                      createdAt:
                        type: string
                        description: Creation timestamp
                      isPostAuthor:
                        type: boolean
                        description: Whether the author is the post author
                      isDeleted:
                        type: boolean
                        description: Is deleted
                      author:
                        required:
                        - handle
                        - name
                        type: object
                        properties:
                          name:
                            type: string
                            description: 유저 이름
                          handle:
                            type: string
                            description: 유저 핸들
                          profileImageUrl:
                            type: string
                            description: 유저 프로필 이미지 URL
                            nullable: true
                        description: Comment author
                      isAccepted:
                        type: boolean
                        description: Is accepted
                      id:
                        type: number
                        description: Comment ID
                      content:
                        type: string
                        description: Comment content
                        nullable: true
                      updatedAt:
                        type: string
                        description: Update timestamp
                    description: Node Content
            pageInfo:
              required:
              - hasNextPage
              - hasPreviousPage
              - size
              type: object
              properties:
                size:
                  type: number
                  description: Page Size
                hasNextPage:
                  type: boolean
                  description: Has Next Page
                hasPreviousPage:
                  type: boolean
                  description: Has Previous Page
                endCursor:
                  type: string
                  description: End Cursor
                  nullable: true
                startCursor:
                  type: string
                  description: Start Cursor
                  nullable: true
              description: Page Info
          description: Comments
...
    CreateCommentRequest:
      title: CreateCommentRequest
      required:
      - content
      type: object
      properties:
        content:
          type: string
          description: Comment content
    CreateCommentResponse:
      title: CreateCommentResponse
      required:
      - id
      type: object
      properties:
        id:
          type: number
          description: Created comment ID
...